backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
tiktoken==0.7.0
mistune==3.0.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
